daemon: Use named sockets to avoid network permission requirement
Flatpak	applications don't work with gvfs if network access is not
granted. This is because GVfs for peer-to-peer communication uses
abstract sockets, which are tied to the network namespace. Let's use
named sockets under /run/user/$UID/gvfsd/ instead, which will allow
applications to use --filesystem=xdg-run/gvfsd to grant access.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/515