google: Set the display name for the root folder also
As per the documentation, it is guaranteed that the display name attribute
is always set. The same applies for the name attribute. However, this is
not true for the root of the google backend. This sometimes causes file
chooser crashes currently. Let's set those attributes also for root.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/523


(cherry picked from commit 731bdbf7bf807c05eddd3a1559c01640d0dfe40e)