smb: Set fast content type independently of other attributes
The G_FILE_ATTRIBUTE_STANDARD_FAST_CONTENT_TYPE attribute is currently
set only if the G_FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE or some other
attributes are requested. Thus it is not set when the fast content type
attribute is requested separately. Let's set the attribute independently
to fix this issue.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/529