client: Skip session bus fallback if operation was cancelled
Recently added warnings when peer-to-peer connection fails (commit 60ca0cea)
makes visible old bug that the session bus fallback is tried regardless of
the fact that the operation was cancelled. Let's stop the operation
immediately when G_IO_ERROR_CANCELLED is returned and don't print the
warnings in this case.