google: Increase number of results in batches for better performance
Currently, only 50 results are returned in one batch, which means that
multiple network requests are needed for folders with a lot files, which
makes it slow. I am not sure there was some reason for this limitation
earlier (e.g. before the cache rework), however, I don't see any
rationals for it currently. Let's increase this to its maximum for
better performance.