dav, http: port to libsoup3
The dav and http backends of the daemon now use libsoup3. Things
should work more or less the same as before.

One caveat is that the soup session no longer defaults to
ssl-strict, so it will no longer silently accept certificates.
This does not affect the DAV backend (it has its own handling
which is replicated in the soup3 port) but it does affect the
http backend.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/208