udisks2: Report unmount progress after showing blocking processes
The "show-unmount-progress" signal isn't emitted when unmounting
is successful while the "show-processes" dialog is open.

The "aborted" signal is sent to close the dialog, not to indicate
an error, and shouldn't be taken into consideration when deciding
whether to emit "show-unmount-progress". Further details on the
bug report.

Get gvfs_udisks2_unmount_notify_stop() to send
"show-unmount-progress" whether there was a prior "aborted" or not.

Closes https://gitlab.gnome.org/GNOME/gvfs/-/issues/546