daemon: Remove connection socket properly
The naming scheme for the non-abstract socket was changed but
`new_connection_data_free` was not adjusted to match.

`GDBusServer` will remove the socket when it stops, but only if
`g_dbus_server_start` was called. So we can simplify the process
somewhat. Also don't bother removing the directory now that all sockets
share it.