daemon: Fix socket owner for gvfsd-admin
Recently, GVfs switched from abstract sockets to named sockets. The
named sockets honors file permissions and thus the socket owner is root
in the case `gvfsd-admin`. This obviously prevents client from connecting.
Let's change the owner to the value of `PKEXEC_UID` to ensure that the
socket is usable by the client.

https://gitlab.gnome.org/GNOME/gvfs/-/issues/552