admin: Pass also XDG_RUNTIME_DIR to pkexec environment
Recently, GVfs switched from abstract sockets to named sockets. The
sockets are created in the `XDG_RUNTIME_DIR` directory. This directory
is not unfortunatelly set for `gvfsd-admin` and thus the root's cache
dir is used instead, which is wrong. Let's pass the `XDG_RUNTIME_DIR`
environment variable to the `pkexec` environment similarly to
`DBUS_SESSION_BUS_ADDRESS`.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/552