ci: Add ipv6 in tags to fix random test suite failures
The DAV	test cases sometimes fails with	"(98)Address already in use:
AH00072: make_sock: could not bind to address 127.0.0.1:8088". This
seems to happen	only on	runners	which don't have ipv6 tag. Let's
temporarily add the tag to fix those failures. It would be best to
programmatically try some range of ports, but this is currently out
of my time resources.