Add PartOf= to all systemd units
The `PartOf` directive sets up a one-way dependency between
`graphical-session.target` and our services, so that when
`graphical-session.target` is stopped or restarted, our services also get
stopped or restarted.

This ensures that gvfs daemons that were started in the context of a GUI
login session (through dbus activation) do not continue to run after the
user has logged out.

See `systemd.unit(5)` for more info.