sftp: Adapt on new OpenSSH password prompts
SFTP backend fails when connecting with the "Timed out when logging in"
error with the recent OpenSSH versions. This is because their password
prompts are now prefixed by the `(user@host)` string, which breaks our
`g_str_has_prefix` checks. Let's use `strstr` instead to fix this issue
and also to not break compatibility with old versions and other vendors.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/584