dav: Rewrite to libsoup async API to fix crashes
Since libsoup3 cannot deal with threads, we cannot use the do_
methods which execute in a thread pool. However, we can implement
these in an async manner, which will bypass the thread pool and
get rid of the issue.

The write methods are left synchronous as they deal with a memory
output stream and do not actually call libsoup. Therefore, we do
not have to care whether they are threaded or not.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/609