dav: Port DNS-SD resolver to async API to fix hangs when mounting
The mount operation hangs for DNS-SD style URIs (i.e. with `dav+sd` and
`davs+sd` schemes) after the recent commit 9203fad5. This is because the
backend now uses the asynchronous APIs, but the `GVfsDnsSdResolver` is still
used synchronously. Let's use asynchronous API for the resolver as well to
fix the hangs.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/618