dav: Drop user from URI as a workaround for Nextcloud bug
Currently, it is not possible to connect to some Nextcloud servers when
a username is part of URI. So this also affects volumes configured over
GOA. The main problem seems to be that Nextcloud doesn't send the
`WWW-Authenticate` header as a part of 401 response in certain cases (e.g.
when 2FA is enabled). A recent libsoup change leads to the situation that
DAV backend doesn't have a chance to specify a password as the `authenticate`
signal is not called at all. This is because libsoup tries to authenticate
with an empty password first. Let's drop the username from URI to avoid	the
password-less login attempt so the backend has at least one chance to try
some password in this case...

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/617