goa: Prevent automounts when resuming from suspension
When on WiFi, the NextCloud account is automatically mounted after
resuming from suspension. This is because the `Account.AttentionNeeded`
property changes its value and GOA volume monitor calls mount operation
in that case. It is not obvious to me why this code is here. Also this
is maybe bug on GOA side. But anyway, other backends simply just unmount
when something changes, but don't mount again automatically. I suppose we
should do the same here as well to avoid these unwanted automounts.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/634