daemon: PATH-expand the sftp backend ssh client
Meson is currently configured to search the gvfs builder's PATH for an
ssh client, and hardcode its fullpath as the canonical ssh client for
the gvfs sftp backend.

This setup breaks in cases where the builder has a different ssh client
from the final runtime root, or where the client's pathes differ.
Builders using OpenEmbedded or buildroot workspaces are particularly
affected.

Instead, set SSH_PROGRAM to `ssh` so that it gets PATH-expanded at
runtime.

Closes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/465
Signed-off-by: Alex Stewart <alex.stewart@ni.com>