ftp: Update ftp->connections when the connection is released
After gvfsd-ftp is connected to the ftp server, if the ftp server
stops, the gvfsd-ftp sending request will be stuck, even after the ftp
server is started, it still cannot be recovered. This is because
ftp->connections is not updated when the connection is released，Thus
let ftp->connections decrement by one when the connection is released.

https://gitlab.gnome.org/GNOME/gvfs/-/merge_requests/149