dav: Prevent usage of NULL when user is not specified
When the username is not part of an URI, then the following message is
printed: `GVFS-CRITICAL **: 12:31:47.663: g_mount_spec_set_with_len_internal:
assertion 'value != NULL' failed`. This is because `g_mount_spec_set` is
called with `NULL`. It possibly lead to crash when it is build with the
`G_DISABLE_CHECKS` option. This bug was introduced recently by the commit
6636d89f. Let's check the value before setting it to fix this.

Related: https://gitlab.gnome.org/GNOME/gvfs/-/issues/644