http: Fill `GFileInfo` with uncompressed sizes
Currently, using `FileProgressCallback` is quite confusing as
`current_num_bytes` is in uncompressed bytes unlike `total_num_bytes`
which is in compressed bytes.

In order to solve this issue, we disable compression in `query_info` by
setting the "Accept-Encoding" flag to `identity` (none). Then whenever
`query_info_on_read` is called, we make sure that the request wasn't set
to accept compresssion and fallback to `query_info` if it is the case.

Fixes #195