gdaemonfile: Disable push/pull for Flatpak
The push/pull methods rely on local paths. This is a problem for Flatpak
applications because the GVfs daemons run outside of it, so the same local
path might refer to a different file. Let's disable the push/pull methods
for Flatpak applications to fix errors when moving/copying.

Related: https://github.com/flathub/org.libreoffice.LibreOffice/issues/23