smb: Allow renaming a file to the same name with a different case
We check to see if the desired file name already exists before doing
a renaming by stat'ing the file (to prevent smb from destroying an
existing file).  Since smb is not case sensitive, the check for an
existing file mistakenly returns true if we are only changing the
filename's case.

Add a second check to see whether we are simply changing the
case of the filename.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/672