trash: Add nfs4 and cifs on list with unreliable monitoring
The nfs4 and cifs filesystems are currently handled as filesystems with
reliable monitoring. However, monitoring on nfs4 and cifs is unreliable,
and thus the trash backend is not aware of all file changes. Let's add
those filesystems on the list with filesystems having unreliable
monitoring in order to ensure that the dir is manually rescanned if needed.