client: Invalidate mount cache
When a client obtains a private connection to a mountable service
identified by dbus_id through dbus-daemon, it may fail because the
mountable has exited. Let's invalidate the corresponding local mount
cache and return a `G_VFS_ERROR_RETRY` error, handle the error in the
appropriate place and retry the requested operation.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/536