ci: Disable Dav test cases
The Dav test cases fails currently. This is because the httpd process
fails with "(98)Address already in use: AH00072: make_sock: could not
bind to address 127.0.0.1:8088" error. It used to work on runners with
the "ipv6" tag, but there is no such currently. It is not clear to me
why it fails, the port looks to be unused. Let's disable the Dav test
cases temporarily to unblock the pipeline.


(cherry picked from commit f84e435dc5ce6af37da10135c1867387ef93e0dc)