sftp/ftp: Ensure that is-symlink is always set
Currently, the `G_FILE_ATTRIBUTE_STANDARD_IS_SYMLINK` attribute is set
only when its value is `TRUE`. This is wrong with the latest changes as
the `GFileInfo created without standard::is-symlink` errors are printed
now from the `g_file_info_get_is_symlink` function among others.
Let's set this attribute also when it is `FALSE`.