network: Don't recompute files when it's already scheduled
When smb-related gsettings keys are changed, the `recompute_files`
function is called directly regardless of the fact that it might be
already scheduled. It is not a big deal as the gesttings keys are not
meant to be changed often, though the rest of the code uses a better
approach. It schedules the `recompute_files` function execution only
if it is not already. Let's use that approach in this case as well.