network: Add support to display SMB files merged
The files provided by the dns-sd backend are displayed merged by default.
The services from the smb browse backend are hidden in the "Windows Network"
folder except those from the default domain (which hasn't worked reliably so
far). Let's add a possiblity to show the "Windows Network" content merged
similar to what is done with the dns-sd files. As a part of this change,
let's don't show the default domain results merged when the "Windows Network"
folder is shown to simplify handling a bit.

Related: https://gitlab.gnome.org/GNOME/gvfs/-/issues/307