smb: Use G_IO_ERROR_NOT_SUPPORTED in case of EXDEV
The smb backend fails with the `G_IO_ERROR_WOULD_RECURSE` error currently
when moving file across filesystem boundaries. This is ok for directories,
but not for files. The copy&delete fallback is not used in that case and
nautilus creates directory instead. Let's return the
`G_IO_ERROR_NOT_SUPPORTED` error instead to fix the mentioned issues.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/702