Revert "smb: Remove redundant variable"
This reverts commit fc3e481634317f455e0c219de1f44b86d1dabeef. It turned
out that the variable is not actually redundant. It seems that the
libsmbclient library can return `EPERM` even without calling the
`auth_callback` in certain cases. In such a case, the `mount_cancelled`
variable can't be set to `TRUE`, so it ends up with a livelock.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/703


(cherry picked from commit 51ed82cad283610ee33b267417602a424af53104)