udisks2: Quote mount paths with g_shell_quote when unmounting
Currently, the unmount operation fails for mount paths that are not
propagated from udisks (e.g. FUSE) and contain non-ASCII characters.
This is because the mount path is escaped by the `g_strescape` function
that drops non-ASCII characters among others. It is not clear why this
function was used here, but it seems right to use the `g_shell_quote`
function instead. Let's do so to fix unmounting in the mentioned cases.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/706