daemonvfs: Reduce sync metadata methods calls
On file delete and moved, we are calling sync methods to update the
metadata.  We don't do anything with the result and there's no
reason we need to wait for this to finish before completing
a move/delete operation.  This is particularly notieable in nautilus
when deleting a folder (therefore recursively calling g_file_delete()).
Use the async version to speed up these operation.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/714