backendtrash: Allow deleting the root folder trash:///
Rather than erroring on delete for the trash's root, use it as a
mechanism to empty the trash.  The only method for emptying the trash
currently is enumerating the top-level, and deleting each top-level
file individually which can have poor performance when there are a
large number of top-level items.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/716