ftp: Fix data connection to IPv6 link-local address
With IPv6, the passive mode data connection address is constructed by
taking the control socket address and replacing the port with the one
specified in the EPSV response. In doing so, we lose the scope-id and
flowinfo properties. The scope-id field in particular is required when
connecting to a link local address, as it determines which link the
address belongs to. Without it, the connection fails.

To fix this, manually construct the GInetSocketAddress instance, copying
all the fields except the port from the old address.