Revert "smbbrowse: Force NT1 protocol version for workgroup support"
This reverts commit 6c8bc39f570ea82cf14e83ce7d1dbdbe569d09d1.

It turned out that the libsmbclient library is able to resolve more names
than the `GResolver` API. So the NT1 protocol may be enforced unnecessarily,
which cause mount failures. GVfs now has WS-Discovery support, so I think
it is perfecly fine to drop this workaround completely. If somebody really
needs workgroup support, it has to tweak protocol versions over the smb.conf
file.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/731