network: Disable usage of smbbrowse backend by default
The `smb:///` location doesn't contain anything unless the `smb.conf`
file is tweaked manually to enforce the NT1 protocol. It also requires
the NT1 support to be enabled on the server side, which is not in the
majority of cases. Let's don't mount the `smb:///` location from the
network backend by default. If somebody needs this anyway, it can be
enabled over gsettings.