smb: Create dummy stat for root dir where access is denied
Due to permission control, some backends will report an error (EACCES)
when getting the root directory attributes, but can access
subdirectories normally. Therefore, we need to construct a dummy stat
for the root directory so that the application (such as gvfsd-fuse) can
smoothly traverse the root directory when accessing subdirectories.