ftp: Preserve mtime when copying to local
Currently, the modification time is not preserved when copying from
FTP mount to the local filesystem. That is because the date is not
explicitly restored in the pull method. However, the modification
time should be preserved by default as it is in other cases. Let's
do it to be consistent.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/346