ftp: Add support for setting mtime attribute
The modification date is not currently preserved when copying to the server.
This is because the FTP protocol doesn't support that. However, there is
a draft of the RFC extension that brings `MFMT` commands for this purpose.
See: https://datatracker.ietf.org/doc/html/draft-somers-ftp-mfxx-04
Let's add support for `MFMT` command as it seems that is supported by servers.

Related: https://gitlab.gnome.org/GNOME/gvfs/-/issues/346