dav: Disable query_info_on_read method
The DAV backend uses the `PROPFIND` request to retrieve file properties
to implement the `query_info` and `enumerate` jobs. However, the
`query_info_on_read` job instead relies on the headers (e.g. `Last-Modified`)
to be set in the reply on the `GET` request. Those headers are not mandatory,
which might cause that results from the `query_info` and `query_info_on_read`
jobs are different. I'm not sure this behavior was intentional, because the
`query_info_on_read` function is not implemented in the DAV backend itself,
but is inherited from the HTTP backend. Let's disable the `query_info_on_read`
job in the DAV backend to get consistent results. This among others fixes
problem with restoring modification time for certain servers.

Fixes: https://gitlab.gnome.org/GNOME/gvfs/-/issues/637