recent: Ignore subdirectory enumeration
The recent backend arbitrarily enumerates the root directory regardless
of the filename that is passed.  This is bad because
1) It returns the incorrect contents of a sudirectory
2) If you attempt to remove a directory from Recents, it will
start an infinite loop of enumerating the root contents

We should NOT, however, enumerate the actual contents of the subdirectories,
because recents are removed via the DELETE command, and if we return
the actual subdirectory contents, that would result in data loss.