onedrive: Support Sharepoint drives
We assumed that that an item parent is always unique, but this is not
the case for remote items (aka shared) which are located on a different
drive. In order to browse those files we need to keep track of the
current item parent and not the one attached to the item in focus.
This way we can enable support for Sharepoint drives.