onedrive: Use names instead of id for monitor events
The onedrive backend uses paths based on file names, not IDs. However,
the file monitor events from the backend are currently emitted with the
paths based on IDs. Consequently, file changes are not reflected by
Nautilus. Let's fix the monitor paths, so the Nautilus is aware of the
file changes.