mtp: Update size when writting
Currently, the `size` member of the RWHandle struct is initialized to an
original size, but it isn't updated when writing. The `seek` operation
uses this member to calculate new offset in case of the `G_SEEK_END` type.
Consequently, the offset might be wrongy calculated. Let's update the
offset when writing to fix seeking.