mtp: Initialize size to 0 when replacing
Currently, the `size` member of the `RWHandle` struct is intialized to
an original file size when replacing. This is wrong, because the file is
truncated. The `seek` operation uses this member to calculate new offset
in case of the `G_SEEK_END` type. Consequently, the offset might be wrongy
calculated. Let's initialized it to 0 to fix seeking.