mtp: Reset offset to file size before writing when appending
Currently, the MTP backend allows to rewrite existing content when
appending. This is possible because seeking is allowed, but the offset
is not reset when writing. The local implementation and other backends
uses the `O_APPEND` flag, or similar. That prevents rewriting the
already existing content. Let's reset the offset to the size manually
when writing to achieve the same behavior.