afp: Reset offset to file size before write when appending
Currently, the AFP backend allows to rewrite existing content when
appending. This is possible because the seek/truncate operation is
allowed, but the offset is not reset when writing. The local
implementation and other backends uses the `O_APPEND` flag, or similar.
Let's reset the offset to the file size manually before write to
achieve the same behavior.