afc: Use AFC_FOPEN_APPEND flag when appending
The AFC backend currently uses the `AFC_FOPEN_RW` flag for the
`afc_file_open` operation to implement the `append_to` vfunc. That
means that it is possible to rewrite existing content. The local
implementation and other backends uses the `O_APPEND` flag, or similar.
That prevents rewriting the already existing content. Let's use the
`AFC_FOPEN_APPEND` to achieve the same behavior.