afp: Use stored size when seeking
Currently, the SEEK_END type requires a blocking call to obtain the real
size except the case when replacing. However, the file size is already
tracked by the AFP backend for the purpose of replace and append operations.
Let's use the stored file size when seeking in all cases to avoid the
blocking when seeking.