afp: Truncate file when replacing
Currently, when directly replacing an existing file with AFP backend,
the file is truncated from the `close_write` operation to the resulting
size. This may cause, however, the parts of the old file can be part
of the new one when seeking to an offset bigger than file size. In that
case, the file should be padded by zero bytes instead. Let's truncate
the file before actually replacing its content to ensure that.