build: Disable google option
This requires libgdata which is unmaintained and the only thing
keeping libsoup2 in GNOME

We recommend that distros keep this feature disabled. If people
miss the Google Drive integration, we suggest that the
Google Drive part of libgdata could be split into a
separate library and ported to libsoup3 for use by gvfs.