mtp: Allow cancelling ongoing folder enumerations
Since we already have support for async enumeration of folders, we might
as well make use of that and check whether our job was cancelled while we
we're getting all the file metadata.

This allows cancelling the ongoing enumeration (which might take a very
long time if there's several thousand files) and therefore using the MTP
device again before the enumeration would finish.

While at it, also use g_autofree for the handlers list.