dav: Add X-OC-Mtime header on push to preserve last modified time
Adds the X-OC-Mtime header to the PUT request when pushing files. While
not a standard part of WebDAV, this is supported by many server
implementations (e.g. Nextcloud), and where not, the header should be
ignored, so that the current behavior does not change. This brings the
WebDAV backend in line with the expected behavior, and the behavior of
other backends, where copying a file typically preserves the last modified
time from source to destination.

Closes #615

Co-authored-by: Ondrej Holy <oholy@redhat.com>
Part-of: <https://gitlab.gnome.org/GNOME/gvfs/-/merge_requests/281>